libports: Port to x86_64
Message-Id: <20230508213136.608575-3-bugaevc@gmail.com>

+++ libports/manage-multithread.c
@@ -229,7 +229,7 @@ ports_manage_port_operations_multithread (struct port_bucket *bucket,
   thread_function (void *arg)
     {
       struct ports_thread thread;
-      int master = (int) arg;
+      int master = (int)(uintptr_t) arg;
       int timeout;
       error_t err;
